--- admin/javascript/src/components/grid-field-action/README.md
@@ -1,13 +1,0 @@
-# GridFieldAction
-
-This component renders a button within a grid-field to handle actions.
-
-## Props
-
-### handleClick (function - required)
-
-The handler for when a button is clicked
-
-#### Arguments
-
- * event - the click event
