--- admin/javascript/src/components/grid-field-action/styles.scss
@@ -1,20 +1,0 @@
-.grid-field-action-component {
-    background: none;
-    border: 0;
-    color: $gray-dark;
-    float: left;
-    padding: 0 8px;
-    height: 52px;
-
-    &:hover {
-        color: darken($gray-dark, 10%);
-    }
-    
-    &::before {
-        font-size: 20px;
-    }
-    
-    &:last-of-type {
-        margin-right: 12px;
-    }
-}
