--- admin/javascript/src/components/grid-field-action/tests/grid-field-action.js
@@ -1,5 +1,0 @@
-jest.dontMock('../index');
-
-describe('GridFieldActionComponent', () => {
-
-});
