--- admin/javascript/src/components/grid-field-cell/README.md
@@ -1,9 +1,0 @@
-# GridFieldCell
-
-This component represents a data cell in a GridFieldRow.
-
-## Props
-
-### width (number)
-
-Set a width relative to the other cells if required. Accepts a number from 1-10 (defaults to 5).
