--- admin/javascript/src/components/grid-field-cell/styles.scss
@@ -1,25 +1,0 @@
-.grid-field-cell-component {
-    display: flex;
-    flex-flow: row nowrap;
-    flex-grow: 1;
-    flex-basis: 0;
-    word-wrap: break-word;
-    overflow-wrap: break-word;
-    word-break: break-word;
-    padding: $grid-y*2 $grid-x*2;
-    line-height: 20px;
-
-    &:first-child {
-        padding-left: $grid-x*2.5;
-    }
-    
-    &:last-child {
-        padding-right: $grid-x*2.5;
-    }
-    
-    @for $i from 1 through 10 {
-        &.cell-width-#{$i} {
-            flex-grow: $i;
-        }
-    }
-}
