--- admin/javascript/src/components/grid-field-cell/tests/grid-field-cell-test.js
@@ -1,5 +1,0 @@
-jest.dontMock('../index');
-
-describe('GridFieldCellComponent', () => {
-
-});
