--- admin/javascript/src/components/grid-field-header-cell/README.md
@@ -1,9 +1,0 @@
-# GridFieldHeaderCell
-
-This component is a cell in a GridFirldHeader component.
-
-## Props
-
-### width (number)
-
-Set a width relative to the other cells if required. Accepts a number from 1-10 (defaults to 5).
