--- admin/javascript/src/components/grid-field-header-cell/tests/grid-field-header-cell-test.js
@@ -1,5 +1,0 @@
-jest.dontMock('../index');
-
-describe('GridFieldHeaderCellComponent', () => {
-
-});
