--- admin/javascript/src/components/grid-field-header/README.md
@@ -1,5 +1,0 @@
-# GridFieldHeader
-
-This component is used to display a tabel header row on a GridFieldComponent.
-
-## Props
