--- admin/javascript/src/components/grid-field-header/styles.scss
@@ -1,3 +1,0 @@
-.grid-field-header-component {
-
-}
