--- admin/javascript/src/components/grid-field-header/tests/grid-field-header-test.js
@@ -1,5 +1,0 @@
-jest.dontMock('../index');
-
-describe('GridFieldHeaderComponent', () => {
-
-});
