--- admin/javascript/src/components/grid-field-row/README.md
@@ -1,9 +1,0 @@
-# GridFieldRow
-
-Represents a row in a GridField.
-
-## Props
-
-### cells (array)
-
-The table data to display in the row.
