--- admin/javascript/src/components/grid-field-row/styles.scss
@@ -1,16 +1,0 @@
-.grid-field-table-component {
-    li.grid-field-row-component {
-        display: flex;
-        flex-flow: row nowrap;
-        width: 100%;
-        border: 0;
-        box-shadow: inset 0 -1px 0 0 $gray-lighter;
-        margin: 0;
-        padding: 0;
-
-        // Header row
-        &:first-child {
-            background: none;
-        }
-    }
-}
