--- admin/javascript/src/components/grid-field-row/tests/grid-field-row-test.js
@@ -1,5 +1,0 @@
-jest.dontMock('../index');
-
-describe('GridFieldRowComponent', () => {
-
-});
