--- admin/javascript/src/components/grid-field-table/README.md
@@ -1,13 +1,0 @@
-# GridFieldTableComponent
-
-This component is used to display structured data in an extendible table layout.
-
-## Props
-
-### Headings (array)
-
-The column headings.
-
-### Rows (array)
-
-The table rows.
