--- admin/javascript/src/components/grid-field-table/styles.scss
@@ -1,5 +1,0 @@
-.grid-field-table-component {
-    display: flex;
-    flex-flow: column nowrap;
-    justify-content: space-between;
-}
