--- admin/javascript/src/sections/grid-field/README.md
@@ -1,3 +1,0 @@
-# GridField
-
-General purpose component for tabular data.
