+++ admin/javascript/src/styles/main.scss
@@ -6,12 +6,7 @@
 /** -----------------------------
  * components
  * ------------------------------ */
-@import "../components/grid-field-table/styles";
-@import "../components/grid-field-cell/styles";
-@import "../components/grid-field-header/styles";
-@import "../components/grid-field-header-cell/styles";
-@import "../components/grid-field-row/styles";
-@import "../components/grid-field-action/styles";
+@import "../components/grid-field/styles";
 @import "../components/north-header/styles";
 @import "../components/north-header-breadcrumbs/styles";
 @import "../components/action/styles";
