+++ gulpfile.js
@@ -194,12 +194,12 @@ gulp.task('bundle-lib', function bundleLib() {
         .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/form',                     { expose: 'form' })
         .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/form-action',              { expose: 'form-action' })
         .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/form-builder',             { expose: 'form-builder' })
-        .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/sections/grid-field',                 { expose: 'grid-field' })
-        .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/grid-field-cell',          { expose: 'grid-field-cell' })
-        .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/grid-field-header',        { expose: 'grid-field-header' })
-        .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/grid-field-header-cell',   { expose: 'grid-field-header-cell' })
-        .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/grid-field-row',           { expose: 'grid-field-row' })
-        .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/grid-field-table',         { expose: 'grid-field-table' })
+        .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/grid-field',               { expose: 'grid-field' })
+        .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/grid-field/cell',          { expose: 'grid-field-cell' })
+        .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/grid-field/header',        { expose: 'grid-field-header' })
+        .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/grid-field/header-cell',   { expose: 'grid-field-header-cell' })
+        .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/grid-field/row',           { expose: 'grid-field-row' })
+        .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/grid-field/table',         { expose: 'grid-field-table' })
         .require(PATHS.ADMIN_JAVASCRIPT_SRC + '/components/hidden-field',             { expose: 'hidden-field' })
         .require(PATHS.FRAMEWORK_JAVASCRIPT_SRC + '/i18n.js',                         { expose: 'i18n' })
         .require(PATHS.FRAMEWORK_JAVASCRIPT_SRC + '/jQuery.js',                       { expose: 'jQuery' })
